Fix printf formats & enable large file support for 32-bit

+++ src/zm_eventstream.cpp
@@ -1097,7 +1097,7 @@ bool EventStream::send_file(const std::string &filepath) {
     Info("File size is zero. Unable to send raw frame %d: %s", curr_frame_id, strerror(errno));
     return false;
   }
-  if (0 > fprintf(stdout, "Content-Length: %jd\r\n\r\n", filestat.st_size)) {
+  if (0 > fprintf(stdout, "Content-Length: %jd\r\n\r\n", static_cast<intmax_t>(filestat.st_size))) {
     fclose(fdj); /* Close the file handle */
     Info("Unable to send raw frame %d: %s", curr_frame_id, strerror(errno));
     return false;
